Refactored code (Removed unused awt.Shape from parser.Shape hierarchy, changed some constants from byte to int due to inconvenience in use and also did some other small changes)

+++ src/main/java/com/itextpdf/text/pdf/parser/PathConstructionRenderInfo.java
@@ -12,39 +12,39 @@ public class PathConstructionRenderInfo {
     /**
      * See {@link com.itextpdf.text.pdf.parser.Path#moveTo(float, float)}
      */
-    public static final byte MOVETO = 1;
+    public static final int MOVETO = 1;
 
     /**
      * See {@link com.itextpdf.text.pdf.parser.Path#lineTo(float, float)}
      */
-    public static final byte LINETO = 2;
+    public static final int LINETO = 2;
 
     /**
      * See {@link com.itextpdf.text.pdf.parser.Path#curveTo(float, float, float, float, float, float)}
      */
-    public static final byte CURVE_123 = 3;
+    public static final int CURVE_123 = 3;
 
     /**
      * See {@link com.itextpdf.text.pdf.parser.Path#curveTo(float, float, float, float)}
      */
-    public static final byte CURVE_23 = 4;
+    public static final int CURVE_23 = 4;
 
     /**
      * See {@link com.itextpdf.text.pdf.parser.Path#curveFromTo(float, float, float, float)}
      */
-    public static final byte CURVE_13 = 5;
+    public static final int CURVE_13 = 5;
 
     /**
      * See {@link com.itextpdf.text.pdf.parser.Path#closeSubpath()}
      */
-    public static final byte CLOSE = 6;
+    public static final int CLOSE = 6;
 
     /**
      * See {@link com.itextpdf.text.pdf.parser.Path#rectangle(float, float, float, float)}
      */
-    public static final byte RECT = 7;
+    public static final int RECT = 7;
 
-    private byte operation;
+    private int operation;
     private List<Float> segmentData;
     private Matrix ctm;
 
@@ -54,23 +54,23 @@ public class PathConstructionRenderInfo {
      *                    E.g. x, y, w, h for rectangle; x, y for line etc.
      * @param ctm         Current transformation matrix.
      */
-    public PathConstructionRenderInfo(byte operation, List<Float> segmentData, Matrix ctm) {
+    public PathConstructionRenderInfo(int operation, List<Float> segmentData, Matrix ctm) {
         this.operation = operation;
         this.segmentData = segmentData;
         this.ctm = ctm;
     }
 
     /**
-     * See {@link #PathConstructionRenderInfo(byte, java.util.List, Matrix)}
+     * See {@link #PathConstructionRenderInfo(int, java.util.List, Matrix)}
      */
-    public PathConstructionRenderInfo(byte operation, Matrix ctm) {
+    public PathConstructionRenderInfo(int operation, Matrix ctm) {
         this(operation, null, ctm);
     }
 
     /**
      * @return construction operation should be performed on the current path.
      */
-    public byte getOperation() {
+    public int getOperation() {
         return operation;
     }
